The Websocket relay used to test RemoteLogix through a web browser
And it works !

Still, this is a VERY basic setup that can easily be reproduced by
a few lines of python, node.js or any language you're accustomed
with.

Listen on port X
Listen on port Y
if message comes from X, send to Y

There's no support for Y to X, since RemoteLogix currently does
not need it.

Signed-off-by: Voyage <voyage@miouyouyou.fr>